Almost done with server represenation of robot and map

+++ main.py
@@ -9,7 +9,10 @@ app = Flask(__name__)
 socketio = SocketIO(app)
 button = 0
 
+
 #Array to hold instances of robots
+startX = 0
+startY = 0
 robotArray = []
 
 #Double Array to represent the warehouse
@@ -34,8 +37,8 @@ def talkToRobot(command):
 
 @socketio.on('connect', namespace='/robot')
 def handleRobotConnect():
-	robotArray.append(Robot(request.sid))
-	warehouse.addRobot(0,0)
+	robotArray.append(Robot(request.sid, startX,startY))
+	warehouse.addRobot(startX,startY)
 	print('Robot ' + robotArray[len(robotArray)-1].getId() + ' joins the warehouse!')
 
 @socketio.on('disconnect', namespace='/robot')
@@ -76,34 +79,32 @@ def handleCommand(msg):
 		return
 
 	c = msg['command']
+	#Temporary ack
 	ack = '1234'
-	#TODO: If move commands: Can robot move without colliding with robots or warehouse walls?
-	# If pick up: It it already holding something?
-	# If drop: Is it holding something?
-	# Update warehouse
 	socketio.emit('move', {'id' : ack, 'command' : c}, namespace= '/robot')
 	send(c, broadcast=True)
 
 	while True:
 		@socketio.on('result')
 		def handleResult(msg):
-			#if(msg['re-id'] == ack):
-			send(msg['status'], broadcast=True)
-			'''
-				r = getRobot(request.sid, robotArray)
-				pos1 = r.getPosition()
-				if(c == 'move-straight'):
-					r.move()
-					pos2 = r.getPosition()
-					warehouse.moveRobot(pos1,pos2)
-					warehouse.showWarehouse()
-				if(c == 'drop'):
-					r.drop()
-				if(c == 'pick up'):
-					r.pickup()
+			if(msg['re-id'] == ack):
+				if(msg['status'] == 'successful'):
+					send(msg['status'], broadcast=True)
+					r = getRobot(request.sid, robotArray)
+					pos1 = r.getPosition()
+					if(c == 'move-straight'):
+						pos2 = r.nextMove()
+						if(warehouse.moveRobot(pos1,pos2)):
+							r.move()
+						warehouse.showWarehouse()
+					if(c == 'drop'):
+						r.drop()
+					if(c == 'pick up'):
+						r.pickup()
+					else:
+						r.turn(c)
 				else:
-					r.turn(c)
-			'''
+					send(msg['status'], broadcast=True)
 		break
 
 
@@ -125,68 +126,8 @@ def video_feed():
 		mimetype='multipart/x-mixed-replace; boundary=frame')
 #TUTORIAL SLUTAR HAR
 '''
-@socketio.on('down')
-def handleDown():
-	if (button == 0):
-		msg = "need to take control first"
-		send(msg, broadcast=True)
-	else:
-		send("down", broadcast=True)
+
 
 if __name__ == '__main__':
-	socketio.run(app, host = "130.243.235.165", debug = True)
-
-'''
-@socketio.on('straight')
-def handleUp():
-	if (button == 0):
-		msg = "need to take control first"
-		send(msg, broadcast=True)
-	else:
-		socketio.emit('move', {'id' : '1234', 'command' : 'move-straight'}, namespace= '/robot')
-		send("move-straight", broadcast=True)
-
-@socketio.on('down')
-def handleDown():
-	if (button == 0):
-		msg = "need to take control first"
-		send(msg, broadcast=True)
-	else:
-		send("down", broadcast=True)
-
-@socketio.on('left')
-def handleLeft():
-	if (button == 0):
-		msg = "need to take control first"
-		send(msg, broadcast=True)
-	else:
-		socketio.emit('move', {'id' : '1234', 'command' : 'move-left'}, namespace= '/robot')
-		send("move-left", broadcast=True)
-
-@socketio.on('right')
-def handleRight():
-	if (button == 0):
-		msg = "need to take control first"
-		send(msg, broadcast=True)
-		return
-	else:
-		socketio.emit('move', {'id' : '1234', 'command' : 'move-right'}, namespace= '/robot')
-		send("move-right", broadcast=True)
-	
-@socketio.on('pick up')
-def handlePickUp():
-	if(button == 0):
-		msg = "need to take control first"
-		send(msg, broadcast=True)
-	else:
-		socketio.emit('move', {'id' : '1234', 'command' : 'pick-object'}, namespace= '/robot')
-
-@socketio.on('drop')
-def handleDrop():
-	if(button == 0):
-		msg = "need to take control first"
-		send(msg, broadcast=True)
-	else:
-		socketio.emit('move', {'id' : '1234', 'command' : 'drop-object'}, namespace= '/robot')
-'''
+	socketio.run(app, host = "192.168.1.106", debug = True)
